requirements.txt
httpx>=0.26.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
